avpl/requirement.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-js-asset==1.2.2
django-mptt==0.12.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
googlemaps==4.4.5
idna==2.10
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
ratelim==0.1.6
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
